djangoapp/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
